# improve completion-related stuff