--- a/src/Entity/AppLsp.hs
+++ b/src/Entity/AppLsp.hs
@@ -21,5 +21,9 @@ lspOptions =
               _save = Just $ InR $ SaveOptions {_includeText = Just False}
             },
       optCompletionTriggerCharacters = Just ['.'],
-      optExecuteCommandCommands = Just [CA.minimizeImportsCommandName]
+      optExecuteCommandCommands =
+        Just
+          [ CA.minimizeImportsCommandName,
+            CA.refreshCacheCommandName
+          ]
     }