--- a/src/Entity/Config/Check.hs
+++ b/src/Entity/Config/Check.hs
@@ -4,5 +4,6 @@ import Entity.Config.Remark qualified as Remark
 
 data Config = Config
   { shouldInsertPadding :: Bool,
+    shouldCheckAllDependencies :: Bool,
     remarkCfg :: Remark.Config
   }